+++ app/Filament/Resources/BotSignalWeightResource/Pages/CreateBotSignalWeight.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\BotSignalWeightResource\Pages;
+
+use App\Filament\Resources\BotSignalWeightResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateBotSignalWeight extends CreateRecord
+{
+    protected static string $resource = BotSignalWeightResource::class;
+}