+++ app/Filament/Resources/BotSignalWeightResource/Pages/EditBotSignalWeight.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\BotSignalWeightResource\Pages;
+
+use App\Filament\Resources\BotSignalWeightResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditBotSignalWeight extends EditRecord
+{
+    protected static string $resource = BotSignalWeightResource::class;
+}